test: extract GenerateCIArtifactName to test helpers
A bit of deduplication can never hurt.
